improvement: support dynamic space in route, including pages for space, proposal and edit
