+++ libs/tailwind.ts
@@ -1,0 +1,3 @@
+export function classNames(...classes) {
+    return classes.filter(Boolean).join(' ')
+}
