+++ tailwind.config.js
@@ -13,7 +13,6 @@ module.exports = {
   },
   plugins: [
     require("flowbite/plugin"),
-    require('@tailwindcss/line-clamp'),
     require('@tailwindcss/aspect-ratio'),
     require('@tailwindcss/forms'),
     require('@tailwindcss/typography')
